requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.75.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Mako==1.2.0
MarkupSafe==2.1.1
minio==7.1.8
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
